+++ backends/carp_webservices/lib/carp_services/carp_service.dart
@@ -122,7 +122,7 @@ class CarpService {
 
     // All other cases are treated as a failed attempt and throws an error
     _authEventController.add(AuthEvent.failed);
-
+    _currentUser = null;
     throw CarpServiceException(
       httpStatus: HTTPStatus(httpStatusCode, response.reasonPhrase),
       message: responseJson["error_description"],
@@ -236,6 +236,7 @@ class CarpService {
 
     // All other cases are treated as a failed attempt and throws an error
     _authEventController.add(AuthEvent.failed);
+    _currentUser = null;
 
     throw CarpServiceException(
       httpStatus: HTTPStatus(httpStatusCode, response.reasonPhrase),
@@ -281,6 +282,9 @@ class CarpService {
       message: responseJson["error_description"],
     );
   }
+
+  /// Logout from CARP
+  Future<bool> logout() => _currentUser = null;
 
   // --------------------------------------------------------------------------
   // USERS
@@ -475,40 +479,6 @@ class CarpService {
     );
   }
 
-  /// Get the informed consent to be shown for this study.
-  ///
-  /// This method return a [RPOrderedTask] which is an ordered list of [RPStep]
-  /// which are shown to the user as the informed consent flow.
-  /// See [research_package](https://pub.dev/packages/research_package) for a
-  /// description on how to create an informed consent in the research package
-  /// domain model.
-  ///
-  /// If there is no informed consent json file on the CARP server, `null` is
-  /// returned.
-  Future<RPOrderedTask> getInformedConsent() async {
-    const String INFORMED_CONSENT_FILE_NAME = 'informed_consent.json';
-
-    info('Downloading informed consent as file : $INFORMED_CONSENT_FILE_NAME');
-    FileStorageReference informedConsentReference =
-        await getFileStorageReferenceByName(INFORMED_CONSENT_FILE_NAME);
-    final File myFile = File("$localFilePath/$INFORMED_CONSENT_FILE_NAME");
-
-    final FileDownloadTask downloadTask =
-        informedConsentReference.download(myFile);
-
-    int response = await downloadTask.onComplete;
-
-    RPOrderedTask informedConsent;
-    if (response == 200) {
-      info('Informed consent downloaded : $INFORMED_CONSENT_FILE_NAME');
-      String jsonString = myFile.readAsStringSync();
-      // TODO - implement when json is implemented....
-      // informedConsent = RPOrderedTask.fromJson(
-      //     json.decode(jsonString) as Map<String, dynamic>);
-    }
-    return informedConsent;
-  }
-
   // --------------------------------------------------------------------------
   // DATA POINT
   // --------------------------------------------------------------------------
@@ -520,19 +490,6 @@ class CarpService {
   // --------------------------------------------------------------------------
   // FILES
   // --------------------------------------------------------------------------
-
-  String _path;
-
-  ///Returns the local path on the device where files can be written.
-  Future<String> get localFilePath async {
-    if (_path == null) {
-      final localApplicationDir = await getApplicationDocumentsDirectory();
-      final directory = await Directory('${localApplicationDir.path}/downloads')
-          .create(recursive: true);
-      _path = directory.path;
-    }
-    return _path;
-  }
 
   /// The URL for the file end point for this [CarpService].
   String get fileEndpointUri =>